+++ pulsar/apps/data/stores/redis/client.py
@@ -169,13 +169,12 @@ class Consumer(pulsar.ProtocolConsumer):
         request = self._request
         if len(request) == 2:
             if response is not False:
-                event = self.event('post_request')
                 if not isinstance(response, Exception):
                     cmnd = request[0][0]
                     response = self.parse_response(response, cmnd, request[1])
                 else:
                     response = ResponseError(response)
-                event.fire(response)
+                self.finished(response)
         else:   # pipeline
             commands, raise_on_error, responses = request
             error = None