+++ pulsar/apps/test/case.py
@@ -1,8 +1,8 @@
 import sys
 import unittest
-import logging
 
-from pulsar import multi_async, coroutine_return
+import pulsar
+from pulsar import coroutine_return
 from pulsar.utils.pep import ispy3k
 from pulsar.apps import tasks
 
@@ -10,7 +10,7 @@ from .utils import TestFunction, TestFailure, is_expected_failure
 
 if ispy3k:
     from unittest import mock
-else:  # pragma nocover
+else:  # pragma     nocover
     try:
         import mock
     except ImportError:
@@ -26,6 +26,9 @@ class Test(tasks.Job):
             testcls = testcls()
         testcls.tag = tag
         testcls.cfg = consumer.worker.cfg
+        testcls._test_suite =  consumer.worker.app
+        testcls._loop = pulsar.get_event_loop()
+        testcls._thread_loop = pulsar.get_thread_loop()
         all_tests = runner.loadTestsFromTestCase(testcls)
         num = all_tests.countTestCases()
         if num:
@@ -50,7 +53,6 @@ class Test(tasks.Job):
           unless the test class should be skipped.
         '''
         cfg = testcls.cfg
-        loop = consumer._loop
         runner.startTestClass(testcls)
         error = None
         sequential = getattr(testcls, '_sequential_execution', cfg.sequential)
@@ -65,7 +67,7 @@ class Test(tasks.Job):
                 yield self.run_test(test, runner, error)
         else:
             all = (self.run_test(test, runner, error) for test in all_tests)
-            yield multi_async(all, loop=loop)
+            yield pulsar.multi_async(all, loop=testcls._thread_loop)
         if not skip_tests:
             yield self._run(runner, testcls, 'tearDownClass', add_err=False)
         runner.stopTestClass(testcls)